third_party/WebKit/LayoutTests/external/wpt/image-decodes/image-decode-svg.html

<!DOCTYPE html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>img.decode()</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel=help href="">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<script>
"use strict";

// src tests
// -------------------
promise_test(function() {
  var img = document.createElementNS('http://www.w3.org/2000/svg', 'image');
  img.setAttributeNS('http://www.w3.org/1999/xlink', 'xlink:href', "/images/green.png");
  return img.decode().then(function(arg) {
    assert_equals(arg, undefined);
  });
}, "(svg) PNG image decodes with undefined.");

promise_test(function() {
  var img = document.createElementNS('http://www.w3.org/2000/svg', 'image');
  img.setAttributeNS('http://www.w3.org/1999/xlink', 'xlink:href',
    "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAIAAA" +
    "D91JpzAAAACXBIWXMAAAsTAAALEwEAmpwYAAAAB3RJTUUH4QUSEioKsyAgyw" +
    "AAABl0RVh0Q29tbWVudABDcmVhdGVkIHdpdGggR0lNUFeBDhcAAAAWSURBVA" +
    "jXY9y3bx8DAwPL58+fGRgYACktBRltLfebAAAAAElFTkSuQmCC");
  return img.decode().then(function(arg) {
    assert_equals(arg, undefined);
  });
}, "(svg) PNG url image decodes with undefined.");

promise_test(function() {
  var img = document.createElementNS('http://www.w3.org/2000/svg', 'image');
  img.setAttributeNS('http://www.w3.org/1999/xlink', 'xlink:href', "/images/green.svg");
  return img.decode().then(function(arg) {
    assert_equals(arg, undefined);
  });
}, "(svg) SVG image decodes with undefined.");

promise_test(function(t) {
  var img = document.createElementNS('http://www.w3.org/2000/svg', 'image');
  img.setAttributeNS('http://www.w3.org/1999/xlink', 'xlink:href', "/non/existent/path.png");
  var promise = img.decode();
  return promise_rejects(t, "EncodingError", promise);
}, "(svg) Non-existent path fails decode.");

promise_test(function(t) {
  var img = document.createElementNS('http://www.w3.org/2000/svg', 'image');
  img.setAttributeNS('http://www.w3.org/1999/xlink', 'xlink:href', "data:image/png;base64,iVBO00PDR0BADBEEF00KGg");
  var promise = img.decode();
  return promise_rejects(t, "EncodingError", promise);
}, "(svg) Corrupt image fails decode.");

promise_test(function(t) {
  var img = document.createElementNS('http://www.w3.org/2000/svg', 'image');
  var promise = img.decode();
  return promise_rejects(t, "EncodingError", promise);
}, "(svg) Path-less image fails decode.");

promise_test(function() {
  var img = document.createElementNS('http://www.w3.org/2000/svg', 'image');
  img.setAttributeNS('http://www.w3.org/1999/xlink', 'xlink:href', "/images/green.png");
  var first_promise = img.decode();
  var second_promise = img.decode();
  assert_true(first_promise !== second_promise);
  return Promise.all([
    first_promise,
    second_promise
  ]);
}, "(svg) Multiple decodes succeed.");

</script>
